refactor(events): move components to homonym folder
