+++ src/components/events/EventCard.vue
