+++ src/components/events/IconDetail.vue
